Commit 3. Changes regarding Michael's comments to the naming convention
