+++ create_diet_class.py
@@ -7,9 +7,9 @@ class CreateDiet(libs.Frame):
         img = libs.Label(self, image=render)
         img.image = render
         img.place(x=0, y=0)
-        self.createwidgets(controller)
+        self.create_widgets(controller)
 
-    def createwidgets(self, controller):
+    def create_widgets(self, controller):
         self.photo = libs.PhotoImage(file="D:\Python\Projekty\Projektb\Dietetyk\Images\strzalka.gif")
-        self.button1 = libs.Button(self, text="Back", bg='white',image = self.photo,
-                         command=lambda: controller.show_frame(libs.StartPageClass.StartPage)).place(x=10,y=10)
+        self.button1 = libs.Button(self, text="Back", bg='white', image = self.photo,
+                                   command=lambda: controller.show_frame(libs.start_page_class.StartPage)).place(x=10, y=10)
