+++ main.py
@@ -1,7 +1,7 @@
 import libs
 
 def main():
-    app = libs.ClassContainer.Classcontainer()
+    app = libs.class_container.ClassContainer()
     app.title(("Dietetyk"))
     app.geometry(("1000x520"))
     app.mainloop()
